+++ front-end-tools/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased changes
 
+- Add `smart-contract-name` dropDown. The `smart-contract-names` are extracted from the module.
 - Add link to developer documentation
 
 ## 1.0.0